+++ variants/file_export.py
@@ -110,7 +110,7 @@ HEADER_FIXED = (
     ("gnomad_oe_lof", "Gnomad constrains lof observed/expected", float),
     ("gnomad_oe_lof_upper", "Gnomad constrains lof observed/expected upper", float),
     ("gnomad_oe_lof_lower", "Gnomad constrains lof observed/expected lower", float),
-    ("pathogenicity_summary_arr", "ClinVar pathogenicity summary", str),
+    ("pathogenicity_arr", "ClinVar pathogenicity summary", str),
 )
 if settings.KIOSK_MODE:
     HEADER_FIXED = tuple(filter(lambda x: not x[0].startswith("inhouse_"), HEADER_FIXED))